backend/requirements.txt
fastapi
uvicorn
sqlalchemy>=2.0.0
aiosqlite
pydantic-settings
google-generativeai
python-dotenv